requirements.txt
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.0
fastapi-cache2==0.2.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
pendulum==3.0.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
time-machine==2.16.0
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.31.0
